requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
brax==0.14.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chex==0.1.91
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dash==3.3.0
dash-bootstrap-components==2.0.4
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.34
dotmap==1.3.30
esquilax==2.1.0
etils==1.13.0
evosax==0.1.6
executing==2.2.1
Farama-Notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.20.2
Flask==3.1.2
flask-cors==6.0.2
flax==0.12.1
fonttools==4.61.0
fprs==1.0.0
fqdn==1.5.1
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.0
humanize==4.14.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==7.1.0
ipympl==0.9.8
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.8.1
jax-tqdm==0.4.0
jaxlib==0.8.1
jaxopt==0.8.5
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jumanji==1.1.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
labmaze==1.0.6
lark==1.3.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.2.4
mistune==3.1.4
ml_collections==1.1.0
ml_dtypes==0.5.4
moojoco==1.1.7
msgpack==1.1.2
mujoco==3.4.0
mujoco-mjx==3.4.0
narwhals==2.14.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.3.5
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.30
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qdax==0.5.0
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.46.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
stack-data==0.6.3
streamlit==1.52.2
tenacity==9.1.2
tensorboardX==2.6.4
tensorflow-probability==0.25.0
tensorstore==0.1.79
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
treescope==0.1.10
trimesh==4.10.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wandb==0.23.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==2.0.1
xmltodict==1.0.2
zipp==3.23.0
